feat(util-credentials): move shared credential utils to util-credential package
