+++ packages/credential-provider-process/package.json
@@ -21,10 +21,10 @@
   },
   "license": "Apache-2.0",
   "dependencies": {
-    "@aws-sdk/credential-provider-ini": "3.20.0",
     "@aws-sdk/property-provider": "3.20.0",
     "@aws-sdk/shared-ini-file-loader": "3.20.0",
     "@aws-sdk/types": "3.20.0",
+    "@aws-sdk/util-credentials": "3.0.0",
     "tslib": "^2.0.0"
   },
   "devDependencies": {
