+++ packages/util-credentials/README.md
@@ -1,0 +1,12 @@
+# @aws-sdk/util-credentials
+
+[![NPM version](https://img.shields.io/npm/v/@aws-sdk/credential-provider-sso/latest.svg)](https://www.npmjs.com/package/@aws-sdk/credential-provider-sso)
+[![NPM downloads](https://img.shields.io/npm/dm/@aws-sdk/credential-provider-sso.svg)](https://www.npmjs.com/package/@aws-sdk/credential-provider-sso)
+
+> An internal package
+
+This package provides shared utilities for all credential providers.
+
+## Usage
+
+You probably shouldn't, at least directly.
