+++ packages/util-credentials/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@aws-sdk/credential-provider-ini",
-  "version": "3.20.0",
-  "description": "AWS credential provider that sources credentials from ~/.aws/credentials and ~/.aws/config",
+  "name": "@aws-sdk/util-credentials",
+  "version": "3.0.0",
+  "description": "Shared utilities for all credential providers.",
   "main": "./dist/cjs/index.js",
   "module": "./dist/es/index.js",
   "scripts": {
@@ -21,13 +21,8 @@
   },
   "license": "Apache-2.0",
   "dependencies": {
-    "@aws-sdk/credential-provider-env": "3.20.0",
-    "@aws-sdk/credential-provider-imds": "3.20.0",
-    "@aws-sdk/credential-provider-web-identity": "3.20.0",
-    "@aws-sdk/property-provider": "3.20.0",
     "@aws-sdk/shared-ini-file-loader": "3.20.0",
     "@aws-sdk/types": "3.20.0",
-    "@aws-sdk/util-credentials": "3.0.0",
     "tslib": "^2.0.0"
   },
   "devDependencies": {
@@ -47,10 +42,10 @@
       ]
     }
   },
-  "homepage": "https://github.com/aws/aws-sdk-js-v3/tree/main/packages/credential-provider-ini",
+  "homepage": "https://github.com/aws/aws-sdk-js-v3/tree/master/packages/util-credentials",
   "repository": {
     "type": "git",
     "url": "https://github.com/aws/aws-sdk-js-v3.git",
-    "directory": "packages/credential-provider-ini"
+    "directory": "packages/util-credentials"
   }
 }
